Refactor ExportDatabaseConfigurationHandler.
